1.这个分支为调试parker电机单轴 型号FL20； 2.修改了SDO的配置 PDO映射也做了修改； 3.整体上比较顺利，需要指出的是电机的正转禁止和反转禁止默认是通过IO点控制的，因此需要将相应IO置位，以打开正反转功能，或者将So-17 So-18参数设为0，使转动限制无效。

+++ mainwindow.h
@@ -70,7 +70,7 @@ private:
     typedef struct
     {
         uint16_t ControlWord;
-        uint8_t ModeCmd;
+        uint16_t ModeCmd;
         int32_t TargetPosi;
         uint16_t PosTorLim;
         uint16_t NegTorLim;
@@ -79,7 +79,7 @@ private:
     {
         uint16_t StatusWord;
         uint16_t LenzeStatus;
-        uint8_t ModeStatus;
+        uint16_t ModeStatus;
         uint16_t ErrorCode;
         int32_t VelValue;
         int16_t TorValue;
